LayoutTests/imported/w3c/web-platform-tests/css/geometry-1/interfaces-expected.txt

CONSOLE MESSAGE: line 7: TypeError: null is not an object (evaluating 'document.getElementById('log').getClientRects')

FAIL Geometry APIs interface IDL tests TypeError: null is not an object (evaluating 'document.getElementById('log').getClientRects')
